% Re-exports of the rstantools generics so users can call
% posterior_predict() / posterior_epred() after `library(brms.deeprv)`.
\name{reexports}
\alias{reexports}
\alias{posterior_predict}
\alias{posterior_epred}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.
\describe{
  \item{rstantools}{\code{\link[rstantools]{posterior_predict}},
    \code{\link[rstantools]{posterior_epred}}}
}
}
\keyword{internal}
